# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kanbalone API
  version: 0.9.18
  summary: Ultra-light local personal kanban API for human and AI collaboration
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:3000
security: []
paths:
...
  /api/boards:
    get:
      summary: List boards
      operationId: listBoards
      responses:
        "200":
          description: Board list
          content:
            application/json:
              schema:
                type: object
                properties:
                  boards:
                    type: array
                    items:
                      $ref: "#/components/schemas/Board"
                required: [boards]
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    post:
      summary: Create board
      operationId: createBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                laneNames:
                  type: array
                  items:
                    type: string
              required: [name]
      responses:
        "201":
          description: Created board detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardDetail"
        "400":
          $ref: "#/components/responses/Error"
  /api/boards/import:
    post:
      summary: Import a board
      description: Supports local board import payloads, including larger seed datasets used for bulk local setup and performance testing.
      operationId: importBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoardExport"
      responses:
        "201":
          description: Imported board detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardDetail"
        "400":
          $ref: "#/components/responses/Error"
  /api/boards/{boardId}:
    get:
      summary: Get board shell
      operationId: getBoard
      parameters:
        - $ref: "#/components/parameters/BoardId"
      responses:
        "200":
          description: Board shell
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardShell"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Rename board
      operationId: updateBoard
      parameters:
        - $ref: "#/components/parameters/BoardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete board and related data
      operationId: deleteBoard
      parameters:
        - $ref: "#/components/parameters/BoardId"
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/boards/{boardId}/events:
    get:
      summary: Subscribe to board updates via SSE
      operationId: subscribeBoardEvents
      parameters:
        - $ref: "#/components/parameters/BoardId"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"boardId":3,"event":"board_updated","sentAt":"2026-04-10T00:00:00.000Z"}
        "404":
          $ref: "#/components/responses/Error"
...
  /api/boards/{boardId}/remote-import:
    post:
      summary: Import a remote issue into a board
      description: Creates one local ticket from a remote issue snapshot. The created ticket uses the remote title and copies the remote body into the local body for initial implementation context. Remote metadata is tracked on the ticket, but board export omits it.
      operationId: importRemoteIssue
      parameters:
        - $ref: "#/components/parameters/BoardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoteIssueImport"
      responses:
        "201":
          description: Imported tracked ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/boards/{boardId}/remote-import/preview:
    post:
      summary: Preview a remote issue before import
      description: Resolves a remote issue without creating a local ticket and reports whether the remote issue has already been imported.
      operationId: previewRemoteIssueImport
      parameters:
        - $ref: "#/components/parameters/BoardId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoteIssueImport"
      responses:
        "200":
          description: Remote issue preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteIssueImportPreview"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/boards/{boardId}/export:
    get:
      summary: Export board
      operationId: exportBoard
      parameters:
        - $ref: "#/components/parameters/BoardId"
      responses:
        "200":
          description: Export payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardExport"
        "404":
          $ref: "#/components/responses/Error"
  /api/tickets/{ticketId}:
    get:
      summary: Get ticket detail
      operationId: getTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "200":
          description: Ticket detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      summary: Update ticket
      operationId: updateTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpsert"
      responses:
        "200":
          description: Updated ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete ticket
      operationId: deleteTicket
      parameters:
        - $ref: "#/components/parameters/TicketId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: integer
...
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: integer
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
            required: [error]
  schemas:
    Board:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, createdAt, updatedAt]
    Lane:
      type: object
      properties:
        id: { type: integer }
        boardId: { type: integer }
        name: { type: string }
        position: { type: integer }
      required: [id, boardId, name, position]
    Tag:
      type: object
      properties:
        id: { type: integer }
        boardId: { type: integer }
        name: { type: string }
        color: { type: string }
      required: [id, boardId, name, color]
    CommentSync:
      type: object
      properties:
        commentId: { type: integer }
        status:
          type: string
          enum: [local_only, pushing, pushed, push_failed]
        remoteCommentId:
          type: [string, "null"]
        pushedAt:
          type: [string, "null"]
          format: date-time
        lastError:
          type: [string, "null"]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [commentId, status, remoteCommentId, pushedAt, lastError, createdAt, updatedAt]
    Comment:
      type: object
      properties:
        id: { type: integer }
        ticketId: { type: integer }
        bodyMarkdown: { type: string }
        bodyHtml: { type: string }
        createdAt: { type: string, format: date-time }
        sync:
          $ref: "#/components/schemas/CommentSync"
      required: [id, ticketId, bodyMarkdown, bodyHtml, createdAt, sync]
...
    TicketRelation:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        laneId: { type: integer }
        isResolved:
          type: boolean
          description: Canonical resolved state. Use this instead of isCompleted.
        isCompleted:
          type: boolean
          deprecated: true
          description: Legacy alias for isResolved. Use isResolved for new clients.
        priority:
          type: integer
          enum: [1, 2, 3, 4]
          description: Ticket priority. 1 = low, 2 = medium, 3 = high, 4 = urgent.
        ref: { type: string }
        shortRef: { type: string }
      required: [id, title, laneId, isResolved, isCompleted, priority, ref, shortRef]
...
    TicketRemote:
      type: object
      properties:
        ticketId: { type: integer }
        provider: { type: string }
        instanceUrl: { type: string }
        resourceType: { type: string }
        projectKey: { type: string }
        issueKey: { type: string }
        displayRef: { type: string }
        url: { type: string }
        title: { type: string }
        bodyMarkdown:
          type: string
          description: Read-only remote issue body snapshot. The ticket bodyMarkdown remains the editable local implementation body.
        bodyHtml:
          type: string
          description: Sanitized HTML rendered from bodyMarkdown.
        state:
          type: [string, "null"]
        remoteUpdatedAt:
          type: [string, "null"]
          format: date-time
        lastSyncedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [ticketId, provider, instanceUrl, resourceType, projectKey, issueKey, displayRef, url, title, bodyMarkdown, bodyHtml, state, remoteUpdatedAt, lastSyncedAt, createdAt, updatedAt]
...
    RemoteIssueImport:
      type: object
      properties:
        provider:
          type: string
          description: "Remote adapter name. Current providers: github, gitlab, redmine."
        laneId:
          type: integer
        instanceUrl:
          type: string
        projectKey:
          type: string
          description: Provider-specific project key, for example owner/repo for GitHub.
        issueKey:
          type: string
          description: Provider-specific issue key, for example issue number for GitHub.
        url:
          type: string
          description: Remote issue URL. For GitHub this can be used instead of projectKey and issueKey.
        postBacklinkComment:
          type: boolean
          description: When true, Kanbalone attempts to post one backlink comment to the remote issue after import. This is opt-in and requires provider comment/write permission.
        backlinkUrl:
          type: string
          description: Optional absolute http/https Kanbalone URL to include in the backlink comment. When omitted, the comment includes only the local ticket reference.
      required: [provider, laneId]
    RemoteIssueImportPreview:
      type: object
      properties:
        provider: { type: string }
        instanceUrl: { type: string }
        resourceType: { type: string }
        projectKey: { type: string }
        issueKey: { type: string }
        displayRef: { type: string }
        url: { type: string }
        title: { type: string }
        state:
          type: [string, "null"]
        remoteUpdatedAt:
          type: [string, "null"]
        duplicate:
          type: boolean
        existingTicketId:
          type: [integer, "null"]
        existingTicketRef:
          type: [string, "null"]
      required:
        - provider
        - instanceUrl
        - resourceType
        - projectKey
        - issueKey
        - displayRef
        - url
        - title
        - state
        - remoteUpdatedAt
        - duplicate
        - existingTicketId
        - existingTicketRef
...
    Ticket:
      type: object
      properties:
        id: { type: integer }
        boardId: { type: integer }
        laneId: { type: integer }
        parentTicketId:
          type: [integer, "null"]
        title: { type: string }
        bodyMarkdown: { type: string }
        bodyHtml: { type: string }
        isResolved:
          type: boolean
          description: Canonical resolved state. Use this instead of isCompleted.
        isCompleted:
          type: boolean
          deprecated: true
          description: Legacy alias for isResolved. Use isResolved for new clients.
        isArchived: { type: boolean }
        priority:
          type: integer
          enum: [1, 2, 3, 4]
          description: Ticket priority. 1 = low, 2 = medium, 3 = high, 4 = urgent.
        position: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        blockerIds:
          type: array
          items:
            type: integer
        blockers:
          type: array
          items:
            $ref: "#/components/schemas/TicketRelation"
        blockedBy:
          type: array
          items:
            $ref: "#/components/schemas/TicketRelation"
        parent:
          oneOf:
            - $ref: "#/components/schemas/TicketRelation"
            - type: "null"
        children:
          type: array
          items:
            $ref: "#/components/schemas/TicketRelation"
        ref: { type: string }
        shortRef: { type: string }
        remote:
          oneOf:
            - $ref: "#/components/schemas/TicketRemote"
            - type: "null"
      required:
        [id, boardId, laneId, parentTicketId, title, bodyMarkdown, bodyHtml, isResolved, isCompleted, isArchived, priority, position, createdAt, updatedAt, tags, comments, blockerIds, blockers, blockedBy, parent, children, ref, shortRef, remote]
...
    TicketUpsert:
      type: object
      properties:
        laneId: { type: integer }
        parentTicketId:
          type: [integer, "null"]
        title: { type: string }
        bodyMarkdown: { type: string }
        isResolved:
          type: boolean
          description: Canonical resolved state. Use this instead of isCompleted.
        isCompleted:
          type: boolean
          deprecated: true
          description: Legacy alias for isResolved. Use isResolved for new clients.
        isArchived: { type: boolean }
        priority:
          type: integer
          enum: [1, 2, 3, 4]
          description: Ticket priority. 1 = low, 2 = medium, 3 = high, 4 = urgent.
        tagIds:
          type: array
          items:
            type: integer
        blockerIds:
          type: [array, "null"]
          items:
            type: integer
    BoardShell:
      type: object
      properties:
        board:
          $ref: "#/components/schemas/Board"
        lanes:
          type: array
          items:
            $ref: "#/components/schemas/Lane"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
      required: [board, lanes, tags]
    BoardDetail:
      allOf:
        - $ref: "#/components/schemas/BoardShell"
        - type: object
          properties:
            tickets:
              type: array
              items:
                $ref: "#/components/schemas/Ticket"
          required: [tickets]
    BoardExport:
      type: object
      properties:
        board:
          $ref: "#/components/schemas/Board"
        lanes:
          type: array
          items:
            $ref: "#/components/schemas/Lane"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        tickets:
          type: array
          items:
            type: object
      required: [board, lanes, tags, tickets]